requirements.txt
tensorflow==2.13.0
fastapi
uvicorn
python-multipart
pillow

matplotlib
numpy